requirements.txt
pytumblr==0.1.2
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0